Config files to setup the Boost dependency automatically using vcpkg
Configuration of vcpkg required is more advance than a normal application install.  It requires git clone, running a bootstrap batch file, and then integrating with a local Visual Studio (I'm not sure how much of this the Visual Studio installer actually does for you - mine didn't put a vcpkg on the path.  Maybe that just needed a call to vcvarsall.bat?):

https://vcpkg.io/en/getting-started